refactor(cli): break up dimos.py into command modules
