+++ dimos/cli/commands/bake.py
@@ -1,0 +1,26 @@
+# Copyright 2026 Dimensional Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""The `dimos bake` command; the implementation lives in dimos.cli.bake."""
+
+from __future__ import annotations
+
+import typer
+
+
+def bake(ctx: typer.Context) -> None:
+    """Compose rust native modules into a single host binary."""
+    from dimos.cli.bake.cli import main as bake_main
+
+    bake_main(list(ctx.args))
