+++ dimos/cli/test_dimos.py
@@ -21,12 +21,9 @@ from typing import Any
 import pytest
 from typer.testing import CliRunner
 
-import dimos.cli.dimos as dimos_cli
-from dimos.cli.dimos import (
-    _normalize_simulation_argv,
-    _with_relay_bridge,
-    main,
-)
+import dimos.cli.commands.lifecycle as lifecycle
+from dimos.cli.commands.lifecycle import _with_relay_bridge
+from dimos.cli.dimos import _normalize_simulation_argv, main
 import dimos.cli.spy.run_spy as run_spy
 import dimos.core.coordination.module_coordinator as module_coordinator
 import dimos.core.coordination.process_lifecycle as process_lifecycle
@@ -220,8 +217,8 @@ def stubbed_run(
     monkeypatch.setattr(run_registry, "cleanup_stale", lambda: 0)
     monkeypatch.setattr(run_registry, "generate_run_id", lambda name: f"test-{name}")
     monkeypatch.setattr(process_lifecycle, "spawn_watchdog", lambda *args, **kwargs: None)
-    monkeypatch.setattr(dimos_cli, "install_signal_handlers", lambda *args, **kwargs: None)
-    monkeypatch.setattr(dimos_cli, "LOG_DIR", tmp_path / "logs")
+    monkeypatch.setattr(lifecycle, "install_signal_handlers", lambda *args, **kwargs: None)
+    monkeypatch.setattr(lifecycle, "LOG_DIR", tmp_path / "logs")
     monkeypatch.setattr(logging_config, "set_run_log_dir", lambda path: None)
     monkeypatch.setattr(get_all_blueprints, "get_by_name_or_exit", blueprints.__getitem__)
     monkeypatch.setattr(
@@ -307,7 +304,7 @@ def test_qualified_global_relay_flag_is_applied_before_composition(
         observed_relay_values.append((global_config.local_relay, global_config.relay_url))
         return blueprint
 
-    monkeypatch.setattr(dimos_cli, "_with_relay_bridge", compose)
+    monkeypatch.setattr(lifecycle, "_with_relay_bridge", compose)
 
     result = CliRunner().invoke(main, ["run", "alpha", "--g.local-relay=true"])
